requirements.txt
python-docx==1.1.2
sentence-transformers==3.2.0
beautifulsoup4==4.13.3
PyMuPDF==1.25.3
regex==2024.11.6
torch==2.5.1
streamlit==1.33.0
PyPDF2==3.0.1
pinecone-client==3.1.0
langchain==0.3.19
langchain-google-genai==2.0.11
python-dotenv==1.0.1